v2 - Directory restructure
1) 디렉토리 구조 재설정 - Component와 Page 분리
