+++ src/components/Detail/Company.js
