+++ src/pages/DetailPage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import DetailContainer from "containers/Detail";
+
+const DetailPage = () => {
+  return <DetailContainer />;
+};
+
+export default DetailPage;
