+++ src/pages/HomePage.js
@@ -1,0 +1,8 @@
+import HomeContainer from "containers/Home";
+import React from "react";
+
+const HomePage = () => {
+  return <HomeContainer />;
+};
+
+export default HomePage;
