+++ src/pages/MoviePage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import MovieContainer from "containers/Movie";
+
+const MoviePage = () => {
+  return <MovieContainer />;
+};
+
+export default MoviePage;
