+++ src/pages/PersonPage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import PersonContainer from "containers/Person";
+
+const PersonPage = () => {
+  return <PersonContainer />;
+};
+
+export default PersonPage;
