+++ src/pages/SearchPage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import SearchContainer from "containers/Search";
+
+const SearchPage = () => {
+  return <SearchContainer />;
+};
+
+export default SearchPage;
