+++ src/pages/TvPage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import TvContainer from "containers/Tv";
+
+const TvPage = () => {
+  return <TvContainer />;
+};
+
+export default TvPage;
